+++ sources/source_json_other_1.py
@@ -2,9 +2,8 @@ import aiohttp
 import json
 import os
 
-from parsers.source_json_other import OtherJsonSource
-
 from schemas.feed import ExplainedFeed
+from sources.source_json_other import OtherJsonSource
 
 
 class OneOtherJsonSource(OtherJsonSource):
