+++ sources/source_json_other_2_prepare.py
@@ -2,10 +2,9 @@ import aiohttp
 import json
 import os
 
-from parsers.source_json_other import OtherJsonSource
-
 from schemas.feed import ExplainedFeed
 from schemas.update import Update
+from sources.source_json_other import OtherJsonSource
 
 
 class PrepareTwoOtherJsonSource(OtherJsonSource):
